src/pyproject.toml
[tool.poetry]
name = "flask-auth"
version = "0.1.0"
description = ""
authors = ["pert <petr777@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask-RESTful = "^0.3.9"
Flask-SQLAlchemy = "^2.5.1"
Flask-JWT-Extended = "^4.4.0"
redis = "^4.3.1"
pydantic = {extras = ["dotenv"], version = "^1.9.1"}
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
